--- src/components/header.css
@@ -1,48 +1,0 @@
-.layout{
-    padding-top: 137px;
-}
-
-.header {
-    position: fixed;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 137px;
-    background-color: #ffffff;
-}
-
-.contents {
-    display: flex;
-    width: 96%;
-    max-width: 1100px;
-    height: 100%;
-    margin: 0 auto;
-    justify-content: space-between;
-}
-
-.img {
-    width: 281px;
-    height: 118px;
-    object-fit: cover;
-}
-
-.navigation {
-    align-items: center;
-    display: flex;
-}
-
-.ul {
-    display: flex;
-    list-style: none;
-}
-
-.menu{
-    margin-left: 15px;
-    margin-right: 15px;
-}
-
-.log {
-    display: flex;
-    align-items: center;
-    margin-left: 250px;
-}
